Endpoints protection

+++ routes/api.php
@@ -14,38 +14,38 @@ Route::middleware(['api'])->group(function () {
     Route::post('/register', [AuthController::class, 'register']);
     Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:sanctum');
 
-    Route::get('/users', [UserController::class, 'getAll']);
-    Route::post('/users', [UserController::class, 'create']);
-    Route::get('/users/{email}', [UserController::class, 'getByEmail']);
-    Route::put('/users/{id}', [UserController::class, 'update']);
-    Route::delete('/users/{id}', [UserController::class, 'delete']);
+    Route::get('/users', [UserController::class, 'getAll'])->middleware('auth:sanctum');
+    Route::post('/users', [UserController::class, 'create'])->middleware('auth:sanctum');
+    Route::get('/users/{email}', [UserController::class, 'getByEmail'])->middleware('auth:sanctum');
+    Route::put('/users/{id}', [UserController::class, 'update'])->middleware('auth:sanctum');
+    Route::delete('/users/{id}', [UserController::class, 'delete'])->middleware('auth:sanctum');
 
-    Route::get('/categories', [CategoriesController::class, 'index']);
-    Route::post('/categories', [CategoriesController::class, 'store']);
-    Route::get('/categories/{id}', [CategoriesController::class, 'show']);
-    Route::put('/categories/{id}', [CategoriesController::class, 'update']);
-    Route::delete('/categories/{id}', [CategoriesController::class, 'destroy']);
+    Route::get('/categories', [CategoriesController::class, 'index'])->middleware('auth:sanctum');
+    Route::post('/categories', [CategoriesController::class, 'store'])->middleware('auth:sanctum');
+    Route::get('/categories/{id}', [CategoriesController::class, 'show'])->middleware('auth:sanctum');
+    Route::put('/categories/{id}', [CategoriesController::class, 'update'])->middleware('auth:sanctum');
+    Route::delete('/categories/{id}', [CategoriesController::class, 'destroy'])->middleware('auth:sanctum');
 
-    Route::get('/suppliers', [SuppliersController::class, 'index']);
-    Route::post('/suppliers', [SuppliersController::class, 'store']);
-    Route::get('/suppliers/{id}', [SuppliersController::class, 'show']);
-    Route::put('/suppliers/{id}', [SuppliersController::class, 'update']);
-    Route::delete('/suppliers/{id}', [SuppliersController::class, 'destroy']);
+    Route::get('/suppliers', [SuppliersController::class, 'index'])->middleware('auth:sanctum');
+    Route::post('/suppliers', [SuppliersController::class, 'store'])->middleware('auth:sanctum');
+    Route::get('/suppliers/{id}', [SuppliersController::class, 'show'])->middleware('auth:sanctum');
+    Route::put('/suppliers/{id}', [SuppliersController::class, 'update'])->middleware('auth:sanctum');
+    Route::delete('/suppliers/{id}', [SuppliersController::class, 'destroy'])->middleware('auth:sanctum');
 
-    Route::get('/products', [ProductController::class, 'getAll']);
-    Route::post('/products', [ProductController::class, 'store']);
-    Route::get('/products/{productName}', [ProductController::class, 'getProductByName']);
-    Route::put('/products', [ProductController::class, 'updateStock']);
-    Route::put('/products/{id}', [ProductController::class, 'updateProduct']);
+    Route::get('/products', [ProductController::class, 'getAll'])->middleware('auth:sanctum');
+    Route::post('/products', [ProductController::class, 'store'])->middleware('auth:sanctum');
+    Route::get('/products/{productName}', [ProductController::class, 'getProductByName'])->middleware('auth:sanctum');
+    Route::put('/products', [ProductController::class, 'updateStock'])->middleware('auth:sanctum');
+    Route::put('/products/{id}', [ProductController::class, 'updateProduct'])->middleware('auth:sanctum');
 
-    Route::get('/clients', [ClientController::class, 'getAllClients']);
-    Route::post('/clients', [ClientController::class, 'store']);
-    Route::get('/clients/{username}', [ClientController::class, 'getClientByUsername']);
-    Route::put('/clients/{id}', [ClientController::class, 'updateClient']);
+    Route::get('/clients', [ClientController::class, 'getAllClients'])->middleware('auth:sanctum');
+    Route::post('/clients', [ClientController::class, 'store'])->middleware('auth:sanctum');
+    Route::get('/clients/{username}', [ClientController::class, 'getClientByUsername'])->middleware('auth:sanctum');
+    Route::put('/clients/{id}', [ClientController::class, 'updateClient'])->middleware('auth:sanctum');
 
-    Route::get('/sales', [SaleController::class, 'getAllSales']);
-    Route::post('/sales', [SaleController::class, 'registerSale']);
-    Route::get('/sales/{id}', [SaleController::class, 'getSaleById']);
-    Route::put('/sales/{id}', [SaleController::class, 'updateSale']);
-    Route::delete('/sales/{id}', [SaleController::class, 'deleteSale']);
+    Route::get('/sales', [SaleController::class, 'getAllSales'])->middleware('auth:sanctum');
+    Route::post('/sales', [SaleController::class, 'registerSale'])->middleware('auth:sanctum');
+    Route::get('/sales/{id}', [SaleController::class, 'getSaleById'])->middleware('auth:sanctum');
+    Route::put('/sales/{id}', [SaleController::class, 'updateSale'])->middleware('auth:sanctum');
+    Route::delete('/sales/{id}', [SaleController::class, 'deleteSale'])->middleware('auth:sanctum');
 });
